build: Add missing include
util-private.h includes glib-i18n-lib.h, which requires GETTEXT_PACKAGE
to be defined. The define comes from config.h,
but that cannot be included in headers, so we have to make sure
that any source file that pulls in util-private.h (or a header
that includes it) includes config.h first.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2566>